unify experience section to about, style footer, rename hero section
