+++ index.html
@@ -5,6 +5,7 @@
     <link rel="icon" href="/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Leonard Zonaphan</title>
+    <link href="/src/assets/styles/main.scss" rel="stylesheet" />
   </head>
   <body>
     <div id="app"></div>
