+++ src/App.vue
@@ -1,18 +1,16 @@
 <script setup>
 import NavigationBar from '@/components/NavigationBar.vue'
 import FooterBar from '@/components/FooterBar.vue'
-import HeroSection from '@/components/HeroSection.vue'
+import HomeSection from '@/components/HomeSection.vue'
 import AboutSection from '@/components/AboutSection.vue'
-import ExperienceSection from '@/components/ExperienceSection.vue'
 import ProjectSection from '@/components/ProjectSection.vue'
 </script>
 
 <template>
   <main>
     <navigation-bar />
-    <hero-section />
+    <home-section />
     <about-section />
-    <experience-section />
     <project-section />
     <footer-bar />
   </main>
