+++ src/assets/styles/main.scss
@@ -1,0 +1,3 @@
+body {
+  font-family: Arial, Helvetica, sans-serif;
+}
